91 - Moves and renames Subscriptions page to the lib with the new name WFOSubscriptionsList page
