+++ packages/orchestrator-ui-components/src/components/WFOSubscriptionsList/types.ts
@@ -1,5 +1,3 @@
-// Todo: preparation for moving the Subscriptions component (table) to the lib
-// https://github.com/workfloworchestrator/orchestrator-ui/issues/149
 export type SubscriptionListItem = {
     subscriptionId: string;
     description: string;
