+++ packages/orchestrator-ui-components/src/pages/subscriptions/index.ts
@@ -1,2 +1,1 @@
-export * from './subscriptionListTabs';
-export * from './types';
+export * from './WFOSubscriptionsListPage';
